pyproject.toml
[project]
name = "ninja-extended"
version = "0.1.0"
description = "Extensions for Django Ninja"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "django-ninja>=1.3.0",
    "pydantic[email]>=2.10.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest-django>=4.9.0",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
    "uv>=0.5.1",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

exclude = ["tests/demo"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    #"D100",   # ignore missing docs
    #"D101",
    #"D102",
    #"D103",
    #"D104",
    #"D105",
    #"D106",
    #"D107",
    #"D200",
    #"D205",
    "D202",
    "D203",
    "D211",
    "D212",
    "D213",
    #"D400",
    #"D401",
    "D413",
    #"D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "ISC001",
    "S101",   # assert
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "DTZ", "PLR"]
